install os packages, including neovim
By having the vim config in a single .vimrc file, it becomes very
portable to work both inside and outwith the devbox.

A single ~/.vimrc file defines the vim plugins, and auto installs them on
first load. There is a single plugin defined to enable a colour scheme.

There is a ~/.config/nvim/init.vim file that simply links to the vim
config, and can be extended in the future with neovim specific config.